Ensure git tags are available for package workflow
Set ``fetch-depth`` to ``0`` for the checkout process in our ``deploy`` job where the package is actually built. We're using _setuptools-scm_ for package versioning; it needs the git tags to be available, and without the full repo history it won't be able to determine a version number for any commits that aren't pushed together with an associated tag.